requirements.txt
# Core HTTP
requests>=2.32.0

# Dashboard
streamlit>=1.35.0
plotly>=5.22.0
pandas>=2.2.0

# Environment
python-dotenv>=1.0.0
